Deselect resources after action modal rendered
Once we you can "select all" (or _bunch_) of instances to perform a
**batch** action on - leaving the user interface in a "selected" state
is less than _ideal_.

This is an effort to enable _safe_ unselect/deselect of all selected
resources.

In doing so, the "modal" calls are using a _"clone" then act_ approach.

I (@lenards) believe this is necessary because the "resources" are
passed as `props` and truly belong to parent-view, <ProjectDetails />.
This means that the collection could be modified out from under the
modals - since it is a reference.

I  believe this is the case; and, explains some odd behavior I saw while
testing out the functionality - but I may be off-base here, so if you
are reading this and think "that `lenards` is totally crazy, he does
not *grok* the JS", find me and set me straight.

A variation of my reasoning and concerns appears in the file:
- troposphere/static/js/modals/instance/resume.js