Fix data loading issue
There was an infinite spinner in DashboardPage. This was strange because
Master listens on all stores and updates everything. It turns out that
Master's children get mounted/rendered before it does. So DashboardPage
mounts/renders and fetches (getAll) from many stores. Then Master mounts and
adds all the listeners, but by now its too late. We wait forever.

To fix the issue I added the appropriate listeners in DashboardPage.
Eventually we will wrap each component in a higher order component which
provides the stores/data it needs.