Fix tropo to pull size information off instance, and not store
Prior an instance's size in details views would be looked up in the size store
which mirroed /api/v2/sizes. When an admin end-dates sizes, the sizes are no
longer at the api and correspondingly in the store resulting in a failed
lookup. Now troposphere just relies on the size information attached to the
instance.