Implement an explicit way to tie data to a component.
Whenever a component depends on data (makes calls to stores to fetch data), we
require the good habit of listening to stores, but nothing enforces that.

Many pages didn't listen to the stores on which they depend. They assume a
populated store. A full page refresh on the dependent page will reveal that
its not subscribed to the data. So the data comes back empty on first blush,
and they don't listen to the store for when the real data does arrive.

See troposphere/static/js/utilities/subscribe.js for some brief docs on the
new way to subscribe to stores in a component.

See components/requests/ResourceHistoryMaster.jsx for an example