Remove unmaintained unit tests
Unfortunately, when Troposphere was moved to webpack for bundling, the
approach for running unit-tests are not moved. This meant that as the
code ended up not being kept up-to-date with changes in components.

Now, this code is removed to avoid confusing anyone coming to the code
base.

The project is focusing providing a BDD suite of tests to ensure that
end-user expectations in behavior are upheld. As that suite grows,
considerations for adding/including fine-grain tests for components
will be revisited.