Change all route manipulating code to appBrowserHistory
For "actions" & "modals", it is difficult (or, perhaps, impossible) to
get a reference to the "history" object associated with the `<Router/>`.

- - -
One approach to this is "don't do the navigation there" (explained more
in the 'note' below).
- - -

The root cause of the bug described in ATMO-1803 [0] was the use of the
`browserHistory` object without the `basename` being set. When a
`basename` is available, it is _prepended_ to all URIs passed in `push`
or `replace` (etc). But in `actions/instance/launch`, we used it and
the resulting page was "invalid" and unhandled by the app.

Here, we introduce a preconfigured instance for the "application's
history", aka `appBrowserHistory`, to ensure that the `basename` is
included.

Note - we might consider that "actions" and "modals", as a _best
practice_, should *not* be navigating in the first place. Then, we
would need to re-architect the usage such that actions provide a promise
that can be handled when resolved (and `done(...)` called. Connor has
quite elegantly shown how this could be done in the Admin UI - and that
could become inspiration.

It's not clear to me at this time if "modals" could benefit from a
similar approach. What is clear though is that "actions" and "modals"
are not the best place for decisions regarding navigation to happen
(though, I am open to hearing "why" they are and "how" we could get
a uniform pattern of _handling_ in place).

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1803