Change argument order to match 'toastr'
Not using the expected order of arguments means that the "verbose" error
appears LARGE and BOLD and the _summary_ appears below it.

In the exposed interface for `toastr`'s method, the "title" is optional.
This means they have opted for the _required_ "message" to be first, &
the "title" to be second in all methods:

see the source links if curious: (error, info, success, warning)
https://github.com/CodeSeven/toastr/blob/2.1.3/toastr.js#L46
https://github.com/CodeSeven/toastr/blob/2.1.3/toastr.js#L68
https://github.com/CodeSeven/toastr/blob/2.1.3/toastr.js#L82
https://github.com/CodeSeven/toastr/blob/2.1.3/toastr.js#L92