Fix invalid variable reference
Failured only exhibited when trying to access Web Desktop action button

It appears that an undefined variable from a previous version of the
`render` methods for <InstanceActionsAndLinks/> was pulled into a
method (maybe an ExtractMethod refactor) and was *not* caught during
review and local testing.

Thanks to @c-mart for catching this.