Track move selected resources actions
This adds two actions to track/observe, one for just clicking on the
"Move selected resources" button and another for the completion of
that action.

The intent here is to see if these two actions appear "as pairs", or
if the community member, after seeing the modal, does not go through
with moving resources.

Note - this is to help inform decisions about "move selected resources"
in the future as a supported interaction.