Ensure help page load data for hyperlinks
When data is finally loaded into the HelpLinkStore, the <HelpPage/>
component is not receiving the call to `updateState`.

This work does a forceUpdate on the page component.

Note: this page is a candidate for being converted to the `subscribe`
connector that Connor (@cdosborn) has written in PR 561 (recently
merge into master).

See ATMO-1844 [0]

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1844