Associate "Move" resources with project-related buttons
This restores the button-bar layout present before Xylotomous-Xenops.

One issue with "Move" and "Remove" buttons (which are operations on the
project, not resources within the project) was that when multiple
resources were select, the "<ResourceActionsButtons/>" was empty and the
buttons move fully to the right. This may be desirable to indicate the
operation is available during "multi", but I believe it is better to
maintain the expected layout and handle any _flow_ improvements during
the re-work of the entire Project Resources view (this is forthcoming).