Define an emulate component & add to the User list
This defines a glyphicon icon that is clicked, allowing someone to find
and emulate an AtmosphereUser easily.

It has a tooltip to help someone discover the icon's _purpose_, and
nicely fit, inline, within the username cell of the table.

This could be "too obscure", and that would degrade the helpfulness.
This will be part of the review.

See ATMO-1602 [0]

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1602