Identify Reboot with power icon
Both "Reboot" and "Redeploy" shared the same _repeat_ icon; here we can
communicate the "power-cycling" of an instance with the "off" icon from
Glyphicon.

This was done how the icon was mapped prior to the following commit:

- 56a7a964171d39d9568f37d07ab509df3e3ebcfa

When the "actionSpec" declarative version of InstanceActionAndLinks was
created, we lost this icon change.

Long live Reboot's icon "off"!