Define route for imaging-request details
Unfortunately I *super* dropped the ball on the Routes within the Admin
tab. I did *not* confirm this functionality of these regarding the
handling of imaging-request. So, apologies are in order </apology>

This defined a "detail" route for `admin/imaging-request` as simply the
`:id` of a request, `admin/imaging-request/:id`. This is a slight
departure from path previously used - I need to confirm/verify this will
not cause problems with support activities.