Handle instance fetch robustly
A `done(..)` will not handle a failure, if the `fetch()` fails for any
reason, then we will not get POLLing dispatched and and subsequent
updates will not be received.

Changing to a `then()`, we have the combination of a done & fail handler
function [0]. When a `fetch()` fails, do the minimal to provide updates:
POLL_INSTANCE.

Also, the `uuid` being set to the `attrs.alias` is necessary. If this is
removed, then the polling calls will be made with an undefined "uuid".

[0] http://api.jquery.com/jQuery.ajax/#jqXHR