Favor builtin Django template filters
There is an existing, available "builtin" filter provided by Django to
convert from newline, `\n`, to break tags, `<br/>`. If we consider the
example below:

```
{{ value | linebreaksbr }}
```

If value is `Atmosphere Notice\nScheduled Downtime`, the output will be
`Atmosphere Notice<br />Scheduled Downtime`. We favor `linebreaksbr` [0]
because it does **not** surround the output with `<p>` tags.

For our usage, we can trust the markup and render it within a template.

Usage then becomes:

```
{{ value | linebreaksbr | safe }}
```

If we does not use `safe`, then `<br/>` tags are escaped.

[0] https://docs.djangoproject.com/en/1.10/ref/templates/builtins/#linebreaksbr