Use stores via subscriptions destructure
The two stores [0] are available from `this.props.subscriptions`, use
ES6/2015 destructuring to _exact_ the stores, by name, and make them
available as local variables.

This commit includes a refactor that, manually, performs an
ExtractMethod [1] to define a TemplateMethod [2], `submitUpdate` that
can be used as the "core" implementation for `approve`, `deny`, and
`re-submit`.

Normally, I would not want to conflate a change with a refactor like
this. I think the elimination in repeated code warrants it - if not,
my abject apologies.

[0] ["ImageRequestStore", "StatusStore"]
[1] https://refactoring.com/catalog/extractMethod.html
[2] https://en.wikipedia.org/wiki/Template_method_pattern