Ensure retrieving HREF from model is safe
Here, we mean "safe" in that an attempt to `.get("href")` will not
result in an error.

An error relates to the help link missing was observed [0]. In this
commit, we favor the "empty" document fragment [1], so this may pop the
to the top of the viewport. But, it will not cause an unhandled error.

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1752
[1] https://www.w3.org/TR/html5/browsers.html#scroll-to-fragid