Remove custom tooltip for react-tooltip
This responds to feedback during review related to using the
`react-tooltip` library over a ReactDOM-style approach. We have
favored this in prior component, but `react-tooltip` appears to
handle nearly all of our usage scenarios.

Thanks to Connor for the feedback and to point out how this component
was used in components/admin/ResourceRequest.jsx