Use consistent hash input for iconsets
The function mapped over the "Instance History Items" is treating the
element passed as `instance`, but it is really _an item_. So when the
hash is computed for the Gravatar icon, it is the "cid" - not the ID
of the instance.

Here, we use the "instance" associated with the _item_ when hashing,
and this resolving the inconsistent iconsets [0]

Consider performing an "argument" rename refactoring on the mapped
function used for rendering "instance histories".

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1848