Use model UUIDs when selecting/unselecting
When using integer identifiers from the database, you can have
unintended "collisions" between identifiers for different models. Here
we leverage the available "alias" identifier to avoid any "collision",
and subsequent "double selection".

Note, these aliases, "uuid" in the model attribuates, are UUIDv4 [1] and
collisions are extremely unlikely, it's "at least one collision is 2.71
quintillion" [2]. We can operate as if there is a guarantee of
uniqueness.

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1320
[1] https://en.wikipedia.org/wiki/Universally_unique_identifier#Version_4_.28random.29
[2] https://en.wikipedia.org/wiki/Universally_unique_identifier#Collisions