Reduce calls to update instance actions
This reduced the potential calls made to `updatedActionsFor` a given
instance. For an instance in a non-final state, we poll for status
changes. That would mean we make a call to `v2/instance/:uuid/actions`
for each poll. This should eliminate at least 50% of the calls to
update instance-actions, and reduces the potential calls by 2/3s.

This change is being made to mitigate the number of calls to the API.