Filter locally rather than call the api
The code in the api is doing too little (comparing if a field is large enough),
it doesn't merit having to load from the api. Since the frontend already has
to reach into the image version to get machines, then machines by provider,
there is not really anything gained by the api call.