Fix method invocation target
In commit 641864f51d4a9005adf8b4ce37c5bfd3e9b4c819, the method here
moved from a private definition to being one available on the prototype
for BaseStore. This code, `fetchOne`, was written against the "private"
implementation - so we update it to use `this` now that the method is
available to all prototypes.

The intent was to only do polling on an instance when it was an in a
non-active state. If it's active, there is no need to keep looking for
in instance.status changes.

Commit 641864f51d4a9005adf8b4ce37c5bfd3e9b4c819 was made while trying
to "shed load" on the API. This change made a significant impact in
that effort.