Allow specific console-log statements for feedback
This is a temporary waiver relating to evaluating these statements in
order to preserve the feedback related to these occurrences. Each one
is trying to hint at a deeper problem in the system or within the data
returned by the API.

We are allowing the occurrences using the eslint-disable & eslint-enable
comment declaratives:

```javascript
/* eslint-disable no-console */
console.warn('Important hint: ' + model);
/* eslint-enable no-console */
```

We should consider replacing these statements with a call out to Raven
or some information gathering abstraction. For a Sentry-specific case,
we could use the following:

```javascript
// [0] see below
Raven.captureMessage('<current-message-in-console-log>');
```

[0] https://raven-js.readthedocs.io/en/1.0.1/usage/index.html#capturing-a-specific-message