Allow "unused variable" involved in data fetch
This is noted in a comment within source, but it could be that the call
to ProfileStore and captured in `username` is done to "eager" fetch data
from that store.

We _could_ just leave this as:
```
stores.ProfileStore.get().get("username");
```

But there may be a better way that will match the level of abstraction
for the component.

**FUTURE**
Also, this might be worth wrapping in the `context` module.