Favor truthy evaluation when assigning
This fixes two violations of `no-extra-boolean-cast` rule within eslint.

It does not appear that we gain anything from using the `!!b`. It is my
understanding that the benefit of "bang-bang" applied to a var, `b`, is
that there is a guarantee that the result will be `typeof` "boolean".

Here, bang-bang is used as part of a condition statement when assigning
to a variable; so there is not a clear _need_ for the condition to be
of type Boolean.