Favor RegEx.test(s) over String.prototype.match(r)
When using String.prototype.match(...), we are given an Array of all the
matches within the String instance. Given the usage of these methods, we
only desire to know _if_ there is a match, we do nothing with the
matches when present. This also means that we need to do some checks to
ensure the String variable is *truthy* to avoid errors calling it. And,
you have an Array returned that you have to cast to get a Boolean; ugh!

You can see more about `.match(...)` here [0].

If we use the `.test(...)` [1] method from RegEx, we now get a boolean
to indicate a match - *and* we don't need to bother with truthy checks.

This turns the two occurrences modified here into 1-liners, which makes
me smile - and it's Friday, so :smile:

[0] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/match
[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/test