Resolve occurrences of "redeclaring" variables
The recommendation for eslint is to enable the rule `no-redeclare`.
The changes within this commit are resolutions to avoid seemingly
unintentional "redeclarations" of variables. Some of these may be
as a result of individuals _expecting_ block scope for JavaScript.

That is a mere guess, but occurrences of violations are often within
new block contexts, like loops or conditionals.