Declaratively ignore isMounted deprecated warnings
We see the following warning as an error now:

```
Waring: isMounted is deprecated. Instead, make sure to clean up
subscriptions and pending requests in componentWillUnmount to prevent
memory leaks.
```

It turns out that the React Team has been against the usage of
isMounted() for some time [0]. Given our custom Flux implementation,
resolving this warning, outright, is not a trivial matter and requires
some thoughtful consideration.

We will be addressing this soon, as we evaluate shifting to Redux or a
"Redux + delta" approach to fetching and handling data.

[0] https://facebook.github.io/react/blog/2015/12/16/ismounted-antipattern.html