Allow `comma-dangle` occurrences
There appears to be more benefits to allowing the comma-dangle then
explicitly banning it.

One of the biggest reasons to allow it would be that as you add to
lists or objects literals, you do not need to modify the preceding line
and this leads to more clear commit 'diff's.

We could enhance the rules to disallow the "dangle" in various usage,
see the the ESLint documentation for the robustness of this rule [0].

[0] http://eslint.org/docs/rules/comma-dangle