Make render of Identity dropdown conditional
This uses `featureFlags.hasProjectSharing()` to determine if the
"form-group" for identity needs to be included in the <ResourcesForm/>.

We have merged Project Sharing (aka User/Group Mapping) into the
mainline to avoid "drift". We know this feature will take multiple
releases before it can be enabled within a production deployment.

Our intend is to make Project Sharing something that can be "toggled"
on, or enabled - then development can enhance the feature without it
falling behind the master branch (e.g. trunk). This concept has been
discussed in a way places [0,1] (usually ThoughtWorks folks [2])

[0] https://trunkbaseddevelopment.com/
[1] https://www.youtube.com/watch?v=lqRQYEHAtpk
[2] like Neal Ford, for example