Fix instance details allocation dropdown
The primary issue is that the paramater `props` was being ignored.
```diff
     getStateFromProps(props) {
-        let { allocationSources, instance } = this.props;
+        let { instance } = props;

-        // This may look strange ========v
-        // This method gets called from  v
-        // getInitialState, when state   v
-        // doesn't exist!                v
-        let { current } = this.state || {};
```

Background:
This component is responsible for rendering the allocation source for a given
instance. When the user selects a new allocation, we patch the instance to
have that allocation, which is a network call, that would result in a small
but noticeable delay in the dropdown.

Solution:

- Make changes to allocation dropdown quick
- Render the actual allocation if our change api call fails

Solution explanation:
Store the allocation locally as allocationId, so that switching an allocation is just a
call to setState and is speedy. However, we also simultaneously fire off the
network request `updateAllocationSource`, which patches an instance with a
new allocation and updates the InstanceStore. This will result in an updated
instance prop, with the new allocation. That triggers getStateFromProps
(above) we use that value (the correct value) to override the allocationId in
state, because it is the source of truth. That ensures that if the newtork
request fails, the dropdown will snap back to the actual value, rather than
the optimistic speedy one.