By default turn off CSS hot reloading
Why:
CSS hot reloading requires the css to be shipped in the js bundle. That means
that our html will show unstyled content, until the js bundle is loaded. This
results in a white page with black text:

    Loading Atmosphere

This will never happen in production, because we ship the css in a link tag
in the head of the document, forcing the css to be applied before the content.

In an effort to bring the developer experience closer to production, we are
making CSS hot reloading opt-in.

In order to enable it, prefix your npm run serve command like so:
```bash
CSS_IN_JS=true npm run serve ...
```

(cherry picked from commit 0d3447b7a9cbb619d2b084c8065d00d8ff4a5f61)