Define utility method to safely build size label
This is another "handle undefined & null" still helper, or utility
method. I think this ought to be a utility given the complexity of the
format for Instance Size(s). I did waffled on whether to destructure
the `size` model to remove the repeated `{$size.attr}` without the
string template literal. It might be a waste of defining local variables
for something like this, so if desired - those could be inlined.