Add wrapping function for Sentry "capture" calls
This is added so that the components are not "bound", or coupled, to the
Raven/Sentry's client. It would act as a facade around the client,
though not all methods are exposed - just "captureMessage" [0]. It does
also allow for additional data to be associated with the captured
occurrence [1].

This could be done like so:

Raven.captureMessage('Increase transparency', {
  extra: { for: 'fun', also: 'dominate' }
});

The second argument will be merged with the context available to Raven
when the message is captured. [2]

[0] https://docs.sentry.io/clients/javascript/usage/#capturing-messages
[1] https://docs.sentry.io/clients/javascript/usage/#passing-additional-data
[2] https://github.com/getsentry/raven-js/blob/864fa390d3f99e872e3b89c779ee10757675d25d/src/raven.js#L398-L405