Define a "wait" indicator for launch operation
This is a special-cased indicator for the scenario where someone is
waiting on "launch" of an instance. The component is an unpolished
placeholder, akin to the "RefreshIndicator" [0], but the idea that
`top` & `left` had to be set meant that (I believe) RefreshIndicator
would require absolute positioning within the footer. This commit opts
for defining a potentially unnecessary component avoid position quirks.

[0] http://www.material-ui.com/#/components/refresh-indicator