Alter model to wait on launch
This introduces the state `waitingOnLaunch` to signal that the modal
should not close until "success" or "failure" occurs.

This leaves the "close" (x) and "CANCEL" elements active so that the
modal _could_ be exited if a requested operation were to "hang" or
become unresponsive to the point was a community member wishing to
abandon the action.

More polish to the animation, text position**, and such are desired.

[0] https://pods.iplantcollaborative.org/jira/browse/ATMO-1153

** the text baseline of "LAUNCHING..." do *not* line up with CANCEL and
it is driving me a bit crazy