Limit Advanced view of Instance Luanch modal
The "Advanced" view is a section with the InstanceLaunchWizardModal
which cannot be viewed while in a `waitingOnLaunch` state. This allows
the modal to "drop" into the "Advanced" section would lead to confusion
with the "state" the modal is in. While `waitingOnLaunch`, we only want
to allow "exit" (via 'close:x' or onCancel).

This continues to honor the flag for disabling the "Advanced" view,
`advancedIsDisabled`.

So, if it's not `advancedIsDisabled` and not `waitingOnLaunch`**, the
"Advanced" view can be shown.

** this is equivalent to !(advancedIsDisabled || waitingOnLaunch) via
DeMorgan's Law, I believe (if I'm wrong - hit me)