Favor coherent label text
This re-purposes the "Cancel" button; yet it accurately describes what
can be done here. The launch has *already* on its' way, we cannot undo
this request via the existing approaches. In this scenario, we use text
that describes the action - "Dismiss".

The launch modal will need to be further explored now that we have the
creation going through api/v2/instances. A complete solution will need
to include an eager saved, "pending" resource to communicate the actual
status of the operation.