Fix error preventing save/add of scripts
Changes related to PR #703 caused an error when using the Advanced
Options of Instance Launch Modal when trying to "Save and Add" a new
script (Deployment Script, aka Boot Script).

The error was three fold:

- the new Deployment type value modeled in this.state
- no handler for `onCreate` was passed, *required*
- no form-element present to choose Deployment type

This area is _hidden_ within the app, but is accessible via an hyperlink
after an Image is selected.

To understand where the error was:

- go to Projects
- select a Project
- Click "NEW > Instance"
- select any Image
- click on "Advanced Options" in bottom, left of modal
- Click on "Create a New Script"

This work resolves the issue and restores the expect flow for creating
new scripts via this "advanced" selection flow.

See also ATMO-2023

[0] https://github.com/cyverse/troposphere/pull/703