Fix boot-script create & add
This corrects an issue introduced when adding deploy type & strategy
field to "boot scripts" (aka deployment scripts). The creation of these
resource(s) failed from the Image Request modal since the fields are
now required.

We could have just "defaulted" the fields to a _reasonable_ choice, but
this introduces the form-input elements to allow choice.