Replace broken isMounted workaround with cancellables approach
I thought that you could track is mounted state in the component and call
setState({ mounted: false }) on componentWillUnmount. That doesn't work,
because setState calls are no-ops in that method. So callback handlers
referring to this.state.mounted (i.e. to determine if the component is still
around) will be wrongfully told that it is still mounted.

The cancellables approach means that you keep track of a list of functions
which you want to cancel on unMount (hence cancellable).