Do not update models when props change
Problem: The views for allocationSource and identities were being reset
whenever any text was typed in the deny-this-request field.

Each of those views allows the user to manipulate a resource. They have to
keep track when the cloud resource changes and when the local resource they're
responsible for changes (i.e. from a user event like typing).

The existing logic said 'reset the local model to the cloud model when a new
cloud model prop comes along'. But props come along even when they don't
change, any re-render of the parent re-renders all children with all props,
not just props that have changed.

This patch works because the only way (now) that local models are changed is
in the success callbacks of the network calls that update the cloud resources.