Remove the admin_url, which is no longer included
This is part of a refactor in the api. The admin_url was submitted to the api,
when a user submitted a resource request so that the RT email would include a
url for staff to click on to manage the resource request. However RT is no
longer going to be used for resource requests so the admin_url is no longer
relevant.