Fix bug where optimistic update prevents api fetch
Problem:
A user submits a resource request from the dashboard. On add, the store
creates this.models to add this optimistic model. Now when the user goes to
visit my-requests, That component will call `getAll` on the resource request
store, which will return the 1 model, and the user won't see what exists in
the api.

Solution:
Change add, so that in addition to adding the optimistic model it fetches from the api.

The solution wasn't as easy as adding a this.fetchModels(), because
fetchModels only fetches when this.models is falsy. After looking at the API I
realized that both getModels and fetchModels do the same thing. There's no way
to say, "Go fetch the damn models". I noticed that every call to fetchModels()
was already guarded like:
```
if (!this.models) { return this.fetchModels() }
```
So the expectation was that fetchModels would indeed fetch the damn models.
There were only a few (obviously-bugridden) places where the call was not
guarded and those were fixed here.