Perform extract-method refactors
This takes the 6 views that make up "Advanced Options" and extracts them
out into a method per view, plus one for Advanced Options.

This also moves around handlers, unifying to the `on${errorName}`
scheme.

Methods that are used to validate are grouped as well.

In general, this is a restructuring but the "extract-method" refactor
was leveraged to try to harmonize the "level of abstraction" each
method operates at.