Fix issue where modals become unscrollable
Problem:
Sometimes when the migrate resources modal appeared it would be unscrollable.

The actual cause is that a modal does cleanup when you call `modal.hide()`.
But it triggers an event and only starts the cleanup on the next event loop --
i.e. modal.hide() is not immediate. If you launch a new modal in the same
stacktrace as calling `.hide()` then the cleanup will get borked.

Solution:
If you want to do something after a modal closes you just listen like so:
```
    $('#myModal').on('hidden.bs.modal', function (e) {
      // do something...
    })

```

This functionality exists already in the BootstrapModalMixin as the
handleHidden prop, so this change just leverages that via promises.

Essentially calling renderModal(...) returns a promise that resolves when the
modal is totally cleaned up. This means that code which wants to render modals
in a sequence can now look like:

```
renderModal(...).then(() => renderModal(...)).then(/*etc*/)
```