Fix admin cannot update another user's identity
Problem:
When an admin would save a quota in the admin resource request view, the save
would fail.

Solution:
Pass a username query paramater in the url

Background:
When the admin clicked 'save', a patch was submitted to /api/v2/identities/.
The response was a 404.
Normally that endpoint only operates the current user's identities. Any
attempt to request or update another users' identities returns a 404. The
current atmosphere implementation expected a query param to be passed:
username. This signals the api to know that a user (must be an admin) wishes
to updates models they don't own.
Because bunching all this functionality into a single end-point doesn't jive
with our models/stores. We created api/v2/admin/ endpoints so that queryparams
don't result in completely different behavior. This hotfix is temporary, and
I'll add an admin endpoint/store for identities, so that the requests to
update/fetch/save are simple.