Alter route varible name (for consistency).
When React-Router populates `this.props.params`, it uses the variable
name from the `<Route />` definition when creating the property names
for `this.props.params`.

Previously, the route was defined as `instances/:id`, which would mean
that `params` is `{id: 11111}` (since integer identifier). This creates
a problem because `<InstanceDetail />` is a child of the Project route
wrapping `<InstanceDetail /`> **and** it is a RouteComponent [0].

This change "imposes" a harmony on the naming to help avoid the
component from having to handle the "differences" in the render.

[0] a component that participates in a Route definition:
```
    <Route path="instances/:instanceId"
    	   component={NewInstanceDetail} />
```