Pass `instance` as arg to render Allocation Source
This uses a helper method to get the instance model associated with an
InstanceHistory "record", or model.

Once the instance is available, we know if this instance belongs to the
user associated with the "context" or another.

This was previously "re-fetched" in renderAllocationSourceSection(),
but this could be simplified if the instance was passed to the method.