Pass instance model as separate `props` value
This view had an unintended violation of the Substitution Principle [0].

The library, React Router, injects `params` onto `this.props` for any
component that participates in a Route definition, we call these "Route
Component". So, by having this view component pass `instance` to the
child view - it creates a situation where the child has to "figure out"
what it has been handed.

Hey, we continued to pass _down_ `params` so it will become available
as `this.props.params`, and it will have an object structure that you'd
expect if injected by React Router:

```
{ instanceId: 11122 }
```

And, we pass `instance` to the child. The component `InstanceDetail`
has two "usage" contexts:

1. as a child of a Project "wrapping context"
2. as a Route Component, only showing Instance-related details

This creates a "coherence" with the usage of `this.props.params`, but
it does *not* eliminate the component from some "logic" checks (e.g.
is `this.props.instance` truthy/available).

This is not optimal, nor perfect. But it makes it so that someone not
hip to the _magic_ of RouteComponents will see a "coherent" use of
params, and it will be "present" in both contexts.

If you disagree, then I offer my sincere apologies. But I think this is
a reasonable approach given the scenario.

Sadly, the errors around this have been present in the application for
more than 5 months (for this, I feel a great shame).

[0] https://en.wikipedia.org/wiki/Liskov_substitution_principle