Listen for changes to MachineRequests
Problem: Machine requests in list view were out of date, when a request was
updated

The ImageMaster renders a list of requests and the current request. When the
current request is either approved/closed/updated etc, we propagated an update
to every component subscribing to the request store. We just needed to listen,
which `subscribe` does for you automatically.