Allow the store to fetch/get a single model
For collections of models we support `fetchModels` and `getAll`, for individual
models we have `fetchModel` and `get`. However the latter methods don't work
like you might hope.

Problem: 1
If you call `get(modelId)` that checks if the model exists and then fetches the
first page of the models from the server. If `modelId` isn't in the first page,
then `get` won't work.

Problem: 2
`fetchModel` actually fetches the modelId from the server, so theres no concern
about pagination, but it sticks its result `this.models`. If this.models
doesn't exist, then you'll get a runtime error. So `fetchModel` only works if
you implicitly have called `getAll` or `fetchModels` first. Not nice.

Solution to: 1
I created a `getModel` analog to `get` which actually fetches the model. I
couldn't replace `get` because the current behavior is used in many places.

Solution to: 2
I fixed `fetchedModels` in a backwards compatible way to prevent the runtime
error. When it fetches a model it adds it to the model cache (like it used to)
and it adds it to `this.fetchedModels`, which is a cache for model instances.
This fixes the implicit need to call store.getAll() as mentioned.