Render an overlay over the update instructions for a resource request
When an admin views a resource request that has already been
approved/closed/denied, require interaction to prevent accidental changing of
resources.

Not a huge fan of this approach, think we can redesign this view to be
simpler, more intuitive, but the primary goal is to make the view more usable.