Pin to React version 15.6.*
The current user interface will likely remain on 15.6.* for the
forseeable future.

Why?

We are working on a new version of Troposphere, codenamed "Kaizen". It
offers a chance to deliver on functionality outlined in 1/DLI [0]. We
have technical debt embedded with the current Troposphere such that a
complete "reset" is both beneficial and attractive to the ability to
rapid delivery new features for Atmosphere.

[0] https://github.com/cyverse/troposphere/wiki/1-DLI