Show deploy scripts as part of Instance details
This adds a <DeployScripts/> resource detail to the instance page,
specifically <InstanceDetailsSection /> so that the "deploy scripts"
included at "launch-time" can be seen by a community members.

It allows the script to be seen, by script title in the details. And,
it gives a "read only" view of the deployment script. This non-editable
view is chosen to avoid someone modifying the script in the model and
potentially expecting the changes to immediately take place on the
active instance.

An "Info" text block could be added to <ScriptShowModal /> above the
<ShowScript /> component that would explain that changes to scripts
will require a "deploy" instance action to be triggered. The block was
omitted in hopes that this PR can be merged without round-trips of
edits on what the text _should_ or *could* say. I love the idea, please
PR if you want to add that.