Hide the "Deleted" resource detail on active VMs
The <EndDate /> component appears with the label "Deleted" on any VM
that is navigated to via the Instance Status History section of the
Dashboard. If a VM (Instance) is Active, it will have an end_date of
null ... and this appears than via moment.js's constructor as an invalid
date.