Define instance action buttons for "batch" mode
This create an "analog" to <InstanceActionButtons ... /> that can handle
multiple instance being selected for deletion. This is a potential
exemplar for how other Instance Actions might act when dealing with a
collection of instances.

Deletion (aka "terminate" or "termination") is the only Instance Action
we can deduce as being _valid_ and _available_ for all selected
instance. Any other action, we would need to have some State Transition
data structure available to reduce (by, perhaps, set intersection) what
possible _actions_ are valid given the desire final-state (aka state
transition).

This component is selected when the feature flag BATCH_RESOURCE_ACTIONS
is set to True.