Define batch-savvy instance delete modal
Note, we do not want to allow Instances with *Attached* Volumes to be
deleted. We can discern that an instance has attached volumes be looking
at the "attachedData" attribute for Volumes within a Project.

The VolumeStore is serving as a quasi-client-side join to help determine
if any of the selected instances *have* attached volumes. Unfortunately,
it appears that we cannot determine/discern this without traversing the
available Volumes within the VolumeStore [*].

This modal is defined with the best "copy text" I could come up with.
There will likely be some desire to tweak the wording to fit the
sensibilities of the community - go forth and English-it-up!

[*] Implementing note - I thought a better way to handle this was to use
the ProjectVolumeStore, since it would narrow the traversal to Volumes
within the _known_ Project. At the time of implementation, it looks like
ProjectVolumeStore is implemented in terms of the VolumeStore. There did
not appear to be any benefit from using this store. Perhaps I'm just not
seeing it - I will offer that this evaluation is likely dimwitted on my
part (I apologize to Sanjay for this).